+++ AlchemistCode/Relics/MidasFruit.cs
@@ -1,6 +1,9 @@
+using System.Collections.Generic;
+using Alchemist.AlchemistCode.Cards.Quest;
 using MegaCrit.Sts2.Core.Commands;
-using MegaCrit.Sts2.Core.Entities.Players;
+using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.Entities.Relics;
+using MegaCrit.Sts2.Core.HoverTips;
 
 namespace Alchemist.AlchemistCode.Relics;
 
@@ -8,16 +11,13 @@ public class MidasFruit : AlchemistRelic
 {
     public override RelicRarity Rarity => RelicRarity.Rare;
 
-    private decimal _goldRemainder;
+    protected override IEnumerable<IHoverTip> ExtraHoverTips => HoverTipFactory.FromCardWithCardHoverTips<UnripeFruit>();
 
-    public override async Task AfterModifyingGoldGained(Player player, decimal amount)
+    public override bool HasUponPickupEffect => true;
+
+    public override async Task AfterObtained()
     {
-        if (player != Owner) return;
-        _goldRemainder += amount;
-        var heals = (int)(_goldRemainder / 15m);
-        if (heals <= 0) return;
-        _goldRemainder -= heals * 15m;
-        Flash();
-        await CreatureCmd.Heal(Owner.Creature, heals);
+        var card = Owner.RunState.CreateCard<UnripeFruit>(Owner);
+        CardCmd.PreviewCardPileAdd([await CardPileCmd.Add(card, PileType.Deck)], 2f);
     }
 }